Split sick days and sick note view controller
