+++ src/main/java/org/synyx/urlaubsverwaltung/sicknote/statistics/SickNoteStatisticsViewController.java
@@ -1,32 +1,18 @@
 package org.synyx.urlaubsverwaltung.sicknote.statistics;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.io.ByteArrayResource;
-import org.springframework.http.ContentDisposition;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.synyx.urlaubsverwaltung.csv.CSVFile;
 import org.synyx.urlaubsverwaltung.person.Person;
 import org.synyx.urlaubsverwaltung.person.PersonService;
-import org.synyx.urlaubsverwaltung.sicknote.sickdays.SickDaysDetailedStatistics;
-import org.synyx.urlaubsverwaltung.sicknote.sickdays.SickNoteDetailedStatisticsCsvExportService;
-import org.synyx.urlaubsverwaltung.web.DateFormatAware;
-import org.synyx.urlaubsverwaltung.web.FilterPeriod;
 
 import java.time.Clock;
-import java.time.LocalDate;
 import java.time.Year;
 import java.time.ZonedDateTime;
-import java.util.List;
-
-import static org.springframework.http.HttpStatus.OK;
 
 /**
  * Controller for statistics of sick notes resp. sick days.
@@ -36,19 +22,13 @@ import static org.springframework.http.HttpStatus.OK;
 class SickNoteStatisticsViewController {
 
     private final SickNoteStatisticsService sickNoteStatisticsService;
-    private final SickNoteDetailedStatisticsCsvExportService sickNoteDetailedStatisticsCsvExportService;
     private final PersonService personService;
-    private final DateFormatAware dateFormatAware;
     private final Clock clock;
 
     @Autowired
-    SickNoteStatisticsViewController(SickNoteStatisticsService sickNoteStatisticsService,
-                                     SickNoteDetailedStatisticsCsvExportService sickNoteDetailedStatisticsCsvExportService,
-                                     PersonService personService, DateFormatAware dateFormatAware, Clock clock) {
+    SickNoteStatisticsViewController(SickNoteStatisticsService sickNoteStatisticsService, PersonService personService, Clock clock) {
         this.sickNoteStatisticsService = sickNoteStatisticsService;
-        this.sickNoteDetailedStatisticsCsvExportService = sickNoteDetailedStatisticsCsvExportService;
         this.personService = personService;
-        this.dateFormatAware = dateFormatAware;
         this.clock = clock;
     }
 
@@ -66,39 +46,10 @@ class SickNoteStatisticsViewController {
         return "thymeleaf/sicknote/sick_notes_statistics";
     }
 
-    @PreAuthorize("hasAnyAuthority('OFFICE', 'SICK_NOTE_VIEW')")
-    @GetMapping("/download")
-    public ResponseEntity<ByteArrayResource> downloadCSV(@RequestParam(value = "from", defaultValue = "") String from,
-                                                         @RequestParam(value = "to", defaultValue = "") String to) {
-
-        final FilterPeriod period = toFilterPeriod(from, to);
-
-        // NOTE: Not supported at the moment
-        if (period.getStartDate().getYear() != period.getEndDate().getYear()) {
-            return ResponseEntity.badRequest().build();
-        }
-
-        final Person signedInUser = personService.getSignedInUser();
-        final List<SickDaysDetailedStatistics> allDetailedSickNotes = sickNoteStatisticsService.getAll(signedInUser, period.getStartDate(), period.getEndDate());
-        final CSVFile csvFile = sickNoteDetailedStatisticsCsvExportService.generateCSV(period, allDetailedSickNotes);
-
-        final HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(new MediaType("text", "csv"));
-        headers.setContentDisposition(ContentDisposition.builder("attachment").filename(csvFile.getFileName()).build());
-
-        return ResponseEntity.status(OK).headers(headers).body(csvFile.getResource());
-    }
-
     private Clock getClockOfRequestedYear(Integer requestedYear) {
         if (requestedYear == null) {
             requestedYear = Year.now(clock).getValue();
         }
         return Clock.fixed(ZonedDateTime.now(clock).withYear(requestedYear).toInstant(), clock.getZone());
-    }
-
-    private FilterPeriod toFilterPeriod(String startDateString, String endDateString) {
-        final LocalDate startDate = dateFormatAware.parse(startDateString).orElse(null);
-        final LocalDate endDate = dateFormatAware.parse(endDateString).orElse(null);
-        return new FilterPeriod(startDate, endDate);
     }
 }
